# Performance Degradation in Derived Datatype creation - Open MPI