@nysal 2016-11-16
@bosilca Can you please review this enhancement?