@bosilca 2016-11-16
With SC on, I didn't had the time to thoughtfully analyze the patch, but I don't think we can skip the c2f index creation for Fortran types. However, instead of always registering the datatype with the c2f array upon datatype creation, we can only do it in the MPI_Type_c2f (with the exception of predefined datatypes).