@ggouaillardet 2016-11-21
@bosilca my understanding is that this patch only optimizes temporary ddt (e.g. they are created, use to build the ddt the user requested, and then deleted), so there is no way this type can be used by the end user (and hence there cannot be a need for the Fortran index)

@AboorvaDevarajan ```OBJ_NEW_CONSTRUCTOR``` is one option, an other option is to declare a new class
(e.g. ```ompi_temp_datatype_t``` "inherits" from ```ompi_datatype_t``` and has the ```__ompi_datatype_temporary_allocate()``` constructor, or the other way around).
as far as i am concerned, that would be more consistent with what we currently do in other places.