@nysal 2016-11-21
@ggouaillardet Thanks for your comments. Yes, the issue we are trying to fix is only for temporary data types. In this case, just the ones used in vector/hvector. I believe there might be a few other places that might benefit too, such as collective algorithms that create a temporary data type (Bruck comes to mind), where we really don't expose the data type to the application. If ompi_temp_datatype_t inherits from ompi_datatype_t, OBJ_NEW will still call all the base class constructors right? If so, then that might not help. We might have to inherit from opal_datatype_t and duplicate ompi_datatype_t struct declaration. We could alternatively make this a little more elegant by making OBJ_NEW a variadic macro and then try to handle it (Not sure we can do this portably). This would also mirror the C++ feature of being able to call different constructors in a class, which is what ompi tries to mimic. Please let us know your comments/suggestions.