@nysal 2016-11-21
Another option is if we can really defer the insertion of the data type to MPI_Type_c2f, like George mentioned. Then that would be a bigger win, as we improve performance for all cases of data type creation.