@nysal 2016-11-21
@ggouaillardet yes, that is a bit more compact. 

@bosilca  We'll take a shot at the MPI_Type_c2f() approach and see how it goes. 
Additionally, I see some potential for MPI_THREAD_MULTIPLE improvements for intrinsic data types in fortran too. Maybe worth exploring that as well. PMPI_Type_f2c calls opal_pointer_array_get_item() in the fast path, which involves taking a lock. If we cache the f2c conversion for predefined data types in a static array, we can avoid the lock for this case.