@ibm-ompi 2016-11-28
Build Failed with GNU compiler! Please review the log, and get in touch if you have questions.

 Gist: https://gist.github.com/ibm-ompi/56872b8a6e4819a04736383bc540bc5a