@bosilca 2017-04-19
A different solution has been integrated into master ( #3369 ).